Change macro/namespace/lib names
I think this better reflects the name of the project. There are and will
be other HEIF plugins for non-Qt things.

Qt seems to claim most things starting with 'Q' or 'q' (and definitely
macros starting with "QT_"), so forgoing the underscores might make it
clearer that this isn't an official Qt plugin.  I could be wrong about
this, though. The identifiers can now also read as "Q the IF image
plugin", which isn't ideal.

Keeping the hyphens where appropriate shouldn't be a problem. There
seems to be more of a convention about that.