Avoid copying decoded data during read (#9)
This fix relies entirely on `heif::Image` using a `shared_ptr` to hold a
reference to its data (or otherwise having a cheap copy/move ctor),
which isn't ideal. However, if that implementation detail were to
change, we would just be in the same situation we were before: creating
a copy of the data.